% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.estimateDisp.par.R
\name{estimateDisp.par}
\alias{estimateDisp.par}
\title{This is an internal function of RUV-III-NB package - parallelized version of \code{edgeR::estimateDisp} function, see \code{help("estimateDisp")} for details.}
\usage{
\method{estimateDisp}{par}(
  y,
  design = NULL,
  group = NULL,
  lib.size = NULL,
  offset = NULL,
  prior.df = NULL,
  trend.method = "locfit",
  tagwise = TRUE,
  span = NULL,
  min.row.sum = 5,
  grid.length = 21,
  grid.range = c(-10, 10),
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1),
  tol = 1e-06,
  weights = NULL,
  BPPARAM = BPPARAM,
  ...
)
}
\arguments{
\item{y}{...}

\item{design}{...}

\item{group}{...}

\item{lib.size}{...}

\item{offset}{...}

\item{prior.df}{...}

\item{trend.method}{...}

\item{tagwise}{...}

\item{span}{...}

\item{min.row.sum}{...}

\item{grid.length}{...}

\item{grid.range}{...}

\item{robust}{...}

\item{winsor.tail.p}{...}

\item{tol}{...}

\item{weights}{...}

\item{BPPARAM}{...}

\item{...}{...}
}
\description{
This is an internal function of RUV-III-NB package - parallelized version of \code{edgeR::estimateDisp} function, see \code{help("estimateDisp")} for details.
}
