% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.extFastruvIIInb_ortho.R
\name{extFastruvIIInb_ortho}
\alias{extFastruvIIInb_ortho}
\title{Removing unwanted variation from matched sequencing count data (with orthogonality constraint)}
\usage{
extFastruvIIInb_ortho(
  Y1,
  Y2,
  M,
  ctl1,
  ctl2,
  ctl3,
  k1 = 1,
  k2 = 1,
  k3 = 1,
  robust = FALSE,
  ortho.W = FALSE,
  lambda.a1 = 0.01,
  lambda.a2 = 0.01,
  lambda.a3 = 0.01,
  lambda.b = 16,
  batch = NULL,
  step.fac = 0.5,
  inner.maxit = 50,
  outer.maxit = 25,
  ncores = 2,
  use.pseudosample = FALSE,
  nc.pool = 20,
  batch.disp = FALSE,
  pCells.touse = 0.2,
  block.size = 5000
)
}
\arguments{
\item{Y1}{raw count as DelayedMatrix object with features from modality one as rows and cells as columns}

\item{Y2}{raw count as DelayedMatrix object with features from modality two as rows and cells as columns}

\item{M}{replicate matrix with number of rows equal to the number of cells and number of columns equal to the number of distinct sub-population of cells: M(i,j) = 1 if cell i is part of sub-population j and 0 otherwise. If a row has all zero entries, it means the corresponding cell is not assumed to belong to any of the known sub-populations (a priori unannotated cells).}

\item{ctl1}{either logical vector of length n with TRUE = controls OR a character vector of names of controls (control features for inferring modality one specific unwanted factors W1)}

\item{ctl2}{either logical vector of length n with TRUE = controls OR a character vector of names of controls (control features for inferring modality two specific unwanted factors W2)}

\item{ctl3}{either logical vector of length n with TRUE = controls OR a character vector of names of controls (control features for inferring joint unwanted factors W3)}

\item{k1}{number of unwanted factors to be estimated for W1. Default is 1.}

\item{k2}{number of unwanted factors to be estimated for W2. Default is 1.}

\item{k3}{number of unwanted factors to be estimated for W3. Default is 1.}

\item{robust}{logical value. Whether to use Huber's robust weight in the iterative reweighted least squares (IRLS) algorithm. Default is FALSE.}

\item{ortho.W}{logical value. Whether the unwanted factors (W) are to be orthogonalized. The default is FALSE.}

\item{lambda.a1}{smoothing parameter for regularizing regression coefficients associated with W1. The default is 0.01.}

\item{lambda.a2}{smoothing parameter for regularizing regression coefficients associated with W2. The default is 0.01.}

\item{lambda.a3}{smoothing parameter for regularizing regression coefficients associated with W3. The default is 0.01.}

\item{lambda.b}{smoothing parameter for regularizing the feature-level intercepts. The default is 16.}

\item{batch}{numeric vector containing batch information for each cell. Must correspond to columns of count matrix. If not supplied, cells are assumed to come from one batch.}

\item{step.fac}{multiplicative factor to decrease IRLS step by when log-likelihood diverges. Default is 0.5}

\item{inner.maxit}{the maximum number of IRLS iteration for estimating mean parameters for a given dispersion parameter. Default is 50}

\item{outer.maxit}{the maximum number of IRLS iteration. Default is 25}

\item{ncores}{The number of cores used for parallel processing. The default number of workers (cores) is 2.}

\item{use.pseudosample}{whether to use pseudocells to define replicates (default is FALSE). Note that the replicates defined by the pseudocells will be used in addition to any replicates defined by the M matrix above. We recommend that use.pseudosample=TRUE be used only for data with UMI.}

\item{nc.pool}{the number of cells per pool (used for defining pseudocells). The default is 20. Only relevant when use.pseudosample=TRUE.}

\item{batch.disp}{whether to fit batch-specific dispersion parameter for each feature. The default is FALSE.}

\item{pCells.touse}{the proportion of a priori annotated cells used to estimate alpha and dispersion parameter (Default=20\%). When pseudocells are used (use.pseudosample=TRUE), ultimately only 10\% of the original subset of cells will be used to estimate alpha.}

\item{block.size}{the maximum number of cells for block processing when estimating W matrix. Larger block size can be quicker but requires higher RAM. Default = 5000.}
}
\value{
A SingleCellExperiment object containing the raw and adjusted data in the assays slot, cell-level data including the estimated unwanted factors and the M matrix in the colData slot and the feature-level data including the coefficients associated with the unwanted factors in rowData slot.
}
\description{
This function performs the extended fast version of ruvIIInb method (with orthogonality constraint) to remove unwanted variation from matched sequencing count data that contains features from two modalities. Currently, only Negative Binomial model for UMI data is supported. It takes raw count matrix with features (e.g., genes/transcript and proteins/ADT in CITE-seq) as rows and samples (cells) as columns.
Users need to specify the set of negative controls (i.e features where the between cells variation is assumed to be solely due to the unwanted variation)
and the (pseudo)replicate matrix that define sets of cells (e.g., group by biology of interest) that can be considered technical replicates.
}
