% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.get.res2.R
\name{get.res2}
\alias{get.res2}
\title{Produce normalized count data after adjusting for unwanted variation for matched sequencing count data}
\usage{
get.res2(
  out,
  type = c("logcounts", "pearson", "quantile", "quantile_spanorm"),
  batch = NULL,
  block.size = 5000
)
}
\arguments{
\item{out}{output of call to extFastruvIIInb_vanilla, extFastruvIIInb_ortho, or extFastruvIIInb_cca function.}

\item{type}{type of normalized data. Supported options are 'quantile' (percentile-adjusted count), ''quantile_spanorm' (the revised percentile-adjusted count), 'pearson' (pearson residuals) and logcounts (log normalised count).}

\item{batch}{numeric vector containing batch information for each sample.Must correspond to columns of count matrix. Only needed if batch-specific dispersion parameter is fitted.}

\item{block.size}{the maximum number of cells for block processing when returning the corrected expression matrix. Larger block size can be quicker but requires higher RAM. Default = 5000.}
}
\value{
A matrix containing the normalized data.
}
\description{
This internal function takes extFastruvIIInb_vanilla, extFastruvIIInb_ortho, or extFastruvIIInb_cca output as input and produce various metrics of normalized data.
Option for metrics of normalized data includes percentile-adjusted count (PAC), the revised percentile-adjusted count (PACnew), Pearson residuals, and log normalised count.
}
