% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.makeSCE2.R
\name{makeSCE2}
\alias{makeSCE2}
\title{Create SingleCellExperiment object containing normalised data for matched sequencing count data}
\usage{
makeSCE2(
  obj,
  cData = NULL,
  batch = NULL,
  assays = c("pearson", "logPAC", "logPAC_new")
)
}
\arguments{
\item{obj}{object containing output of call to ruvIII.nb or fastruvIII.nb function.}

\item{cData}{A data frame containing cell-level metadata. This data frame will be used as 'colData' in the SingleCellExperiment object and must be specified.}

\item{batch}{numeric vector containing batch information for each sample. Must correspond to columns of count matrix. Only needed if batch-specific dispersion parameter is fitted.}

\item{assays}{type of corrected data to return. Available types are "logcounts","logPAC" and "pearson".}
}
\value{
A SingleCellExperiment object with normalized data added to the 'assays' slot. The log percentile-adjusted count is stored in the 'logPAC' component of the 'assays' slot, the revised logPAC is in the 'logPAC_new' component, and the Pearson residuals is in the 'pearson' component.
}
\description{
This function takes extFastruvIIInb_vanilla, extFastruvIIInb_ortho, or extFastruvIIInb_cca output as input and creates a SingleCellExperiment (SCE) object containing various metrics of normalised data.
}
